+++ packages/lending/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @naviprotocol/lending
 
+## 2.0.5
+
+### Patch Changes
+
+- fix address-balance coin selection
+
 ## 2.0.4
 
 ### Patch Changes